// web/src/generated/models/ConnectionConfig.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ConnectionConfig = {
    created_at?: string;
    description?: string;
    enabled?: boolean;
    id?: string;
    is_default?: boolean;
    name?: string;
    updated_at?: string;
    url?: string;
};